+++ sdk/src/driftClient.ts
@@ -1813,17 +1813,24 @@ export class DriftClient {
 	 * Get the associated token address for the given spot market
 	 * @param marketIndex
 	 * @param useNative
+	 * @param tokenProgram
 	 */
 	public async getAssociatedTokenAccount(
 		marketIndex: number,
-		useNative = true
+		useNative = true,
+		tokenProgram = TOKEN_PROGRAM_ID
 	): Promise<PublicKey> {
 		const spotMarket = this.getSpotMarketAccount(marketIndex);
 		if (useNative && spotMarket.mint.equals(WRAPPED_SOL_MINT)) {
 			return this.wallet.publicKey;
 		}
 		const mint = spotMarket.mint;
-		return await getAssociatedTokenAddress(mint, this.wallet.publicKey);
+		return await getAssociatedTokenAddress(
+			mint,
+			this.wallet.publicKey,
+			undefined,
+			tokenProgram
+		);
 	}
 
 	public createAssociatedTokenAccountIdempotentInstruction(
@@ -4250,17 +4257,17 @@ export class DriftClient {
 
 		const preInstructions = [];
 		if (!outAssociatedTokenAccount) {
+			const tokenProgram = this.getTokenProgramForSpotMarket(outMarket);
 			outAssociatedTokenAccount = await this.getAssociatedTokenAccount(
 				outMarket.marketIndex,
-				false
+				false,
+				tokenProgram
 			);
 
 			const accountInfo = await this.connection.getAccountInfo(
 				outAssociatedTokenAccount
 			);
 			if (!accountInfo) {
-				const tokenProgram = this.getTokenProgramForSpotMarket(outMarket);
-
 				preInstructions.push(
 					this.createAssociatedTokenAccountIdempotentInstruction(
 						outAssociatedTokenAccount,
@@ -4274,17 +4281,17 @@ export class DriftClient {
 		}
 
 		if (!inAssociatedTokenAccount) {
+			const tokenProgram = this.getTokenProgramForSpotMarket(outMarket);
 			inAssociatedTokenAccount = await this.getAssociatedTokenAccount(
 				inMarket.marketIndex,
-				false
+				false,
+				tokenProgram
 			);
 
 			const accountInfo = await this.connection.getAccountInfo(
 				inAssociatedTokenAccount
 			);
 			if (!accountInfo) {
-				const tokenProgram = this.getTokenProgramForSpotMarket(outMarket);
-
 				preInstructions.push(
 					this.createAssociatedTokenAccountIdempotentInstruction(
 						inAssociatedTokenAccount,
@@ -4390,9 +4397,11 @@ export class DriftClient {
 
 		const preInstructions = [];
 		if (!outAssociatedTokenAccount) {
+			const tokenProgram = this.getTokenProgramForSpotMarket(outMarket);
 			outAssociatedTokenAccount = await this.getAssociatedTokenAccount(
 				outMarket.marketIndex,
-				false
+				false,
+				tokenProgram
 			);
 
 			const accountInfo = await this.connection.getAccountInfo(
@@ -4404,16 +4413,19 @@ export class DriftClient {
 						outAssociatedTokenAccount,
 						this.provider.wallet.publicKey,
 						this.provider.wallet.publicKey,
-						outMarket.mint
+						outMarket.mint,
+						tokenProgram
 					)
 				);
 			}
 		}
 
 		if (!inAssociatedTokenAccount) {
+			const tokenProgram = this.getTokenProgramForSpotMarket(inMarket);
 			inAssociatedTokenAccount = await this.getAssociatedTokenAccount(
 				inMarket.marketIndex,
-				false
+				false,
+				tokenProgram
 			);
 
 			const accountInfo = await this.connection.getAccountInfo(
@@ -4425,7 +4437,8 @@ export class DriftClient {
 						inAssociatedTokenAccount,
 						this.provider.wallet.publicKey,
 						this.provider.wallet.publicKey,
-						inMarket.mint
+						inMarket.mint,
+						tokenProgram
 					)
 				);
 			}